Migration protocol/ping_spec to rspec
